requirements.txt
Flask==1.1.4
Flask-Session==0.4.0
Flask-Caching==1.10.1
flask-restx==0.4.0
python-dotenv==0.18.0